api/files: Split endpoints in files and folders controller + added directory resolver trait
